Conflicts:
	assembler/src/debruijn/graph_construction.hpp
	assembler/src/debruijn/graph_simplification.hpp
	assembler/src/debruijn/graphio.hpp
	assembler/src/debruijn/simplification.cpp
	assembler/src/debruijn/stats/debruijn_stats.hpp
	assembler/src/include/de/insert_size_refiner.hpp

+++ assembler/src/debruijn/genomic_info_filler.cpp
@@ -91,7 +91,7 @@ void GenomicInfoFiller::run(conj_graph_pack &gp, const char*) {
         CovModel.Fit();
 
         gp.ginfo.set_genome_size(CovModel.GetGenomeSize());
-        gp.ginfo.set_ec_bound(CovModel.GetErrorThreshold());
+        gp.ginfo.set_ec_bound((double)CovModel.GetErrorThreshold());
         // ginfo.set.trusted_bound(CovModel.GetLowThreshold());
     }
     INFO("EC coverage threshold value was calculated as " << gp.ginfo.ec_bound());
